pyproject.toml
[build-system]
requires = ["setuptools>=42", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "recmodel"
version = "v1.0"
description = "IC recommendation model"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "pandas==1.5.3",
    "numpy==1.24.3",
    "pyspark==3.4.0",
    "scikit-learn==1.2.2",
    "loguru==0.7.0",
    "pyyaml==6.0",
    "gputil==1.4.0",
    "pydantic==1.10.9",
    "qcore==1.9.0",
    "unidecode==1.3.6",
    "typer==0.9.0",
    "tqdm==4.64.1",
    "pyarrow==11.0.0",
    "torch==2.0.1",
    "pytorch-lightning==2.0.3",
    "apache-airflow==2.6.1",
    "checksumdir==1.2.0",
    "xxhash==3.3.0",
    "bayesian-optimization==1.4.3",
    "matplotlib==3.8.0",
    "faiss-gpu==1.7.2",
    "pendulum==2.1.2",
    "plotly==5.18.0",
    "types-requests==2.31.0.20240106",
    "types-pytz==2023.3.1.1",
    "connexion==2.14.2",
    "xlrd==2.0.1",
    "flask-session==0.5.0",
    "mlflow==2.16.0",
    "fastshap@git+https://github.com/AnotherSamWilson/fastshap.git",
    "cudf-cu12==23.8.0",
    "cuml-cu12==23.8.0",
    "onnxscript==0.1.0.dev20241101"]

[project.optional-dependencies]
dev = [
    # Add the list from requirements.txt here
    "black==23.3.0",
    "flake8==6.0.0",
    "mypy==1.8.0",
    "isort==5.12.0",
    "types-PyYAML==6.0.12.10",
    "pre-commit==3.3.2",
    "pytest==7.3.2"
]
test = [
    # Add the list from requirements.txt here
    "pytest==7.3.2",
]

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]
